Removed styled components from personal project
