--- src/components/PersonalProject.tsx
@@ -1,76 +1,0 @@
-import * as React from 'react';
-import styled from 'styled-components';
-import { IStyledProps } from '../utils/theme';
-
-export interface IPersonalProject {
-  name: string;
-  description: string;
-  url: string;
-  noteOnUrl?: string;
-  sourceUrl?: string;
-}
-
-interface IPersonalProjectProps {
-  project: IPersonalProject;
-}
-
-const StyledDiv = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  margin-top: 1.5rem;
-  border-top: 3.5px solid
-    ${(props: IStyledProps) => props.theme.colors.secondary};
-  border-radius: 5px;
-  background: ${(props: IStyledProps) => props.theme.colors.card};
-  .body,
-  .footer {
-    padding: 1rem;
-  }
-  .footer {
-    border-bottom-left-radius: 5px;
-    border-bottom-right-radius: 5px;
-    background: ${(props: IStyledProps) =>
-      props.theme.colors.backgroundLighter};
-  }
-  h3 {
-    margin-top: 0;
-    margin-bottom: 0;
-  }
-`;
-
-export const PersonalProject: (props: IPersonalProjectProps) => JSX.Element = ({
-  project,
-}) => {
-  return (
-    <StyledDiv>
-      <div className="body">
-				{project.url ? (
-					<a href={project.url} target="_blank" rel="noopener noreferrer">
-          	<h3>{project.name}</h3>
-        	</a>
-				) : (
-					<h3>{project.name}</h3>
-				)}
-        <p>{project.description}</p>
-      </div>
-      <div className="footer">
-				{project.url ?
-					<a href={project.url} target="_blank" rel="noopener noreferrer">
-						See project {project.noteOnUrl && `(${project.noteOnUrl})`}
-					</a> :
-					<br />
-				}
-        {project.url && project.sourceUrl && <br />}
-				{project.sourceUrl ?
-					<a href={project.sourceUrl} target="_blank" rel="noopener noreferrer">
-						Source on Github
-					</a> :
-					<br />
-				}
-      </div>
-    </StyledDiv>
-  );
-};
-
-export default PersonalProject;
