+++ src/components/personal-projects/personal-projects.module.scss
@@ -1,0 +1,16 @@
+.PersonalProjectsGrid {
+  margin-top: 2.5rem;
+
+  h2 {
+    text-align: center;
+    text-decoration: underline;
+
+    @media (min-width: 768px) {
+      grid-column: 1 / span 2;
+    }
+
+    @media (min-width: 1200px) {
+      grid-column: 1 / span 4;
+    }
+  }
+}
